removed opening of browser at startup
If the browser is set to remember previous sessions, you'll end up with
duplicates to remove every time you initialize the server. Since you can
still find the option in the notification area it seems better to let
the user choose when to open the page and when to just silently start
the service. In the future we might want to enable another way to select
the option to start at startup.