updated formatting and fixed saving edgecase
If you saved a lot rapidly while the server was off, it'd add the saving
requests on top of each other causing some edgecases with buggy
behavour. It'd always recover when the server was on and it saved
sucsessfully a few times, but adding the check in bookmarkManager.js
should prevent this. The machine i tested it on is too fast for me to
properly stack and test it though, so some more testing might be needed.

Formatting is just for consistency, now that i'm enforcing 4 it works
more nicely with my editor and all files will remain consistent to this
standard which will help prevent confusion.